WIP make `cumulative_txo_count` part of the Block header
This will greatly simplify some post-processing stuff that happens
to the blockchain, it will mean that a bunch of services that
we have don't have to process the entire blockchain in order to
get these counts, which will become important when the blockchain
gets big.

I would like to add some tests also that this gives a correct
and valid way of computing global indices of transactions, that
agrees with the way we were doing it before this way existed.