Add a pre-push hook, prevents accidentally pushing branches to origin
When we set up mobilecoin repo, some engineers wanted that we prevent
new branches from being created for pull requests in the main repo
itself, they want everyone to use forks, and this makes some tools
work better.

Initially we tried to make github itself reject such pushes but
it was found that there is no way to do it.

Instead, this pre-push hook can be installed locally to prevent
accidentally pushing to origin, and suggest pushing to fork and
opening a PR from your fork instead.