Move ReprBytes32 to its own crate and generalize it (#79)
* Move ReprBytes32 to its own crate and generalize it

We currently have a lot of code that is like manually grabbing
`prost::encoding::*;` and using it to create custom serializations
with prost.

This is a hazard, because that is not part of the public-facing
interface of prost and it is not documented. It also changes
periodically as upstream refactors the library. The only reason
these symbols are publicly visible at all is that there is no way
that `prost-derive` can function if the proc-macro generated code
cannot access these symbols. When we grab `prost::encodings::*;`
it is creating a maintanence hazard that will make it very hard
ever to uprev prost again.

In the past, every cryptographic primitive that went into `prost`
or `serde` required custom bindings code, which was very long.
ReprBytes32 was a quick stopgap to prevent the code duplication
and take us down a route that would be more maintainable.

It turns out that we have some types that are not exactly 32
bytes, and then we have started grabbing the `prost::encodings::*`
again to support those.

Separately from this, we have had a problem in `keys` crate where
we need to be able to get the bytes of any `Kex` public key.
Previously, `Kex` was only implemented for `X25519` and was not
implemented for `Ristretto`, which prevented us from writing generic
(curve agnostic) code in a lot of places.

In https://github.com/mobilecoinofficial/mobilecoin/pull/74 we fixed
this, but we needed to introduce a trait `GetBytes` as a stopgap.

I propose that `GetBytes` can go away and be replaced with this
version of `ReprBytes`. I propose that `ReprBytes32` can also go
away and be replaced with this version of `ReprBytes`.

I believe that `ReprBytes` should potentially be moved out of
mobilecoin repo and treated as a "fundamental trait" in the RustCrypto
ecosystem. Then, crates like `prost` and `serde` would integrate with
`ReprBytes` trait, and e.g. `Dalek` crates would not have to know
about `prost` or `Dalek` would not have to know about `prost` or
`serde`, they would only have to know about `ReprBytes`.

Even if RustCrypto folks don't want to take ownership of this,
this will reduce a lot of nasty code duplication in our repo now,
and may help to clean up the keys crate, and make it able to achieve
its goals of abstracting both X25519 and Ristretto.

* Use re-exports for prost and serde, per review comments

* Add feature-gating, fix Cargo.lock per review comments