Add fog_authority_key_fingerprint to AccountKey, sig to PublicAddress
This just adds fields to the structures, it doesn't actually compute
signatures or give them the right values, or check the signatures.

It also adds version numbers to both structs. This gives us an
escape if we have to make backwards incompatible changes to the way
that public addresses are handled, fog reports are looked up, change
the interpretations of some of these fields, etc.

This also removes Serde serialization of public address, which was
only being used in the keyfile. It makes public address implement
prost instead.

This also removes external.RistrettoPublic from the proto for
public address. It replaces it with CompressedRistretto and adds
some more missing conversions. This leads to small changes in mobilecoind.

This also updates the b58-payloads format to include the version number
and the fog authority sig. I believe that this change is backwards
compatible for any payload not containing fog data.

We should possible delete RistrettoPublic message in external.proto,
I believe that it should never be used.