Add a url-encoding scheme for public addresses and payment requests
This is an alternative to the b58-encoding scheme, outlined by James
sometime in Jan or Feb.

The idea is to build on the fog url, encoding other elements of
the public address into either the path or query parameters of the
fog url. This makes the url understandable to humans roughly,
since you can see the amount and the memo without b58 decoding it,
and see whether it is a fog address, if it has a fog signature, etc.

There is a description of the motivation and goals in the README.md

This also adds test coverage.

This also tests that unicode characters round trip