New fields in account keys and public addresses (#240)
* Add fog_authority_key_fingerprint to AccountKey, sig to PublicAddress

This just adds fields to the structures, it doesn't actually compute
signatures or give them the right values, or check the signatures.

It also adds version numbers to both structs. This gives us an
escape if we have to make backwards incompatible changes to the way
that public addresses are handled, fog reports are looked up, change
the interpretations of some of these fields, etc.

This also removes Serde serialization of public address, which was
only being used in the keyfile. It makes public address implement
prost instead.

This also removes external.RistrettoPublic from the proto for
public address. It replaces it with CompressedRistretto and adds
some more missing conversions. This leads to small changes in mobilecoind.

This also updates the b58-payloads format to include the version number
and the fog authority sig. I believe that this change is backwards
compatible for any payload not containing fog data.

We should possible delete RistrettoPublic message in external.proto,
I believe that it should never be used.

* address review comments, code comments

* Remove version field per review, add fog_report_key field per discussion

This also makes the b58-payloads validate that fog-url actually parses
according to the mc-util-uri::FogUri object

* Rename fog_url -> fog_report_url, make a conversion in mobilecoind-json

Per review comments

* Address more review comments, removing pub, rename to fog_report_url

* Address more review comments, reorder authority_sig and report_key

* Reorder arguemnts to b58-payloads constructors per previous commit

* Fix use statements order