Remove external.RistrettoPublic, replace with CompressedRistretto
Generally, we should not use RistrettoPublic in serialized types,
we should try to use CompressedRistretto, because decompressing
a ristretto point requires elliptic curve operations and is pretty
slow. In early performance profiling, consensus spent most of its
time doing this, and that is why we created the CompressedRistretto
type.

The protobuf type external.RistrettoPublic is actually identical
to external.CompressedRistretto, we don't need both, and we should
discourage people from using RistrettoPublic in their serialized
structs.