Move blake2b_256 out of transaction crate to a new crate
Matt and I discussed this in slack a few months ago but I never
got around to doing this. The point is that this is not specific
to mobilecoin transactions, other things may want to use it without
relying on the entire transaction crate, which contains a lot of
stuff.

At present I'm doing this because this is needed for our Schnorrkel
wrapper, but I don't want the Schnorrkel wrapper to depend on the
transaction crate.