Used fixed64 in transaction amount
(1) Using variable size fields in transaction data means that the
protobuf size reflects something about the numbers in the data,
or which transaction it is, even in encrypted form.
(2) This data is uniformly distributed 64 bit numbers, but the varint
encoding is optimized for small numbers, and is less efficient for
uniformly distributed u64's than the fixed64 encoding. It pays on the
representation size of the large u64's to make the small u64 values
have a shorter representation. But that isn't beneficial for this
data.