Make slog actually work inside of sgx enclave/impl crates
When we planned out the way that we build enclaves, one of the main
questions was "how do we unit test the mission critical code that
lives in enclaves". We split this logic into `_impl` crates that
can compile as std, without having any sgx-only bits inside them.

Currently, none of these crates support logging, because the
`mc-common::logger` module is std-only. We built an `mc-sgx-slog`
interface which sends slog structures logs over an OCALL to untrusted
to be logged on the global logger. But because this code was
enclave-specific, there was no way to actually use this in the
enclave/impl crates without breaking the tests. And so `mc-sgx-slog`
has been basically unused. None of the `_impl` crates use it, and none
of the tests even use the `[test_with_logger]` functionality, because
they can't.

After this revision we make the following changes:

- `mc-common` depends unconditionally on `slog`, and configures it
  in a `no_std` mode.
- `mc-common` unconditionally re-exports `mc-common::logger`, giving
  the `mc-common::logger::log` and `mc-common::logger::slog` parts,
  because these are portable to the enclaves.
- Everything that depends on `slog` now gets it from the `mc-common`
  re-export instead of duplicating the dependency in its cargo toml.
- `slog` is patched to fix minor problems in its no_std build when
  `dynamic_keys` feature is turned on.
- Functionality in `mc-common::logger` that depends on `slog_scope`
  is gated on `log` feature as before, and this requires `std`.
- Functionality in `mc-common::logger` that depends on `loggers`
  feature is still gated on `loggers` as before.
- The `mc-sgx-slog::untrusted` module is moved to the `mc-sgx-urts`
  crate with the other logging related stuff. This also means that
  some times that a server needs to depend `mc-sgx-slog` "only for linkage"
  are removed, because it gets the needed code from `mc-sgx-urts` now.
- `mc-sgx-urts` now sends `eprintln` and `panic` messages to
  `global_log` instead of `eprintln`.
- `consensus-enclave-impl` owns a `Logger` object where it sends log
   messages
- Tests in `consensus-enclave-impl` use `test_with_logger` and build
  the enclave using the test logger, to demonstrate that the functionality
  works.

The biggest drawback in this commit is that it makes the `mc-common::logger`
module messy, but this module was already very messy and was re-exporting
a lot of complicated and variously portable stuff from `slog-*` crates.

We propose to consider changes to the API of the `mc-common::logger` module
in another commit, because that will be onerous and will require touching
almost the entire project, anything that uses the logger. By scoping this
commit so that we don't change the API of `mc-common` we make it possible
to land it and unblock other work that needs logging in enclaves.

It seems that the logger module needs to be redesigned with the idea that
there are three different components that different consumers need:
- `slog` itself, which is totally portable
- `slog_scope`, which is not portable to enclave, depends on `std` and a
  quite complex `ArcSwap` crate. This is used to manage the lifetime
  of the global logger instance. It also depends on thread local vectors of
  logger pointers, to manage the scope loggers. This stuff is no_std
  compatible only in nightly compiler. But we already have several
  pieces of code that are nightly only. IMO We should consider whether
  to patch this crate to make it `no_std` compatible, which would
  simplify stuff somewhat. We would have to get rid of ArcSwap but
  that might be for the best anyways, it's difficult for me to see
  why we need all that, we could require that `set_global_logger`
  is only called once and not allow swapping out global logger on
  the fly. But it's not for sure that we even need
  the `global_log` API in enclaves, we can probably get by without it.
- `sentry` etc., which depends on `std` and an http stack, which we
  cannot compile with `libmobilecoin` or for the phones.