Redo change to make Amount proto object have a fixed size (#391)
This is needed in some backend services, because serialized TxOut
objects will then have a fixed size, and so we won't leak anything
about them through the size of the buffer representing them.

This is a replay of mobilecoin #333, which was reverted in #336
due to concerns that it would break testnet compatibility unless
we made a way to migrate the ledger. The idea we had was that
we would do it right before launch once we were no longer making
testnet releases.