Address NCC group cryptobox comments
There were four pieces of feedback:

(1) There was a subtle cryptogram malleability issue, where a
cryptogram could be modified to produce a new cryptogram which
would decrypt to the same message, but for a different private key.
It would be very difficult to exploit this without knowing the
private key that the original message targetted, but it is not very
hard to prevent this source of malleability.
(2) Use hkdf-expand step, not only the hkdf-extract step.
(3) Fix a buffer overflow in fixed-buffer
(4) Create FOG-182 ticket to track version negotiation in cryptobox

This commit addresses the first 3 issues, doesn't do anything with
version negotation, which we aren't really doing right now, the client
just always uses the latest version.

This commit also increases crypto-box major version to 1, to create
an explicit versioning message if new and old servers/clients are
mixed across this revision.