WIP MCC-959 side-channel resistant merkle proof validation
The previous merkle proof has the following structure:

- index (indicates which number leaf this TxOut is in the merkle tree)
- highest_index (indicates how many TxOuts are in the merkle tree at time of proof)
- several TxOutMembershipElements, which include a hash corresponding to
  an internal node in the tree, and a range explaining which leafs
  are covered by this hash

The previous merkle proof validation routine assumes no particular structure
on the membership elements, it sorts them, throws them into hash maps,
etc. and as a result it is complicated and leaks information about the
path that the merkle proof is following over side channels. Because the
path determines the entire merkle proof, this means that we leak the
whole thing, in the SGX security model.

This PR does the following:

- Introduce a `compose_adjacent_membership_elements` function
  which takes two membership elements, checks if their ranges are
  adjacent in the tree, and computes the hash of their parent in
  the tree if so. This uses subtle constant time equality tests
  and conditional assignments to ensure that there are no branches
  at all, except leading to erroring code paths.
- Merkle proof validation then works strictly from the bottom up,
  using `try_fold` to fold successive membership elements from
  the membership proof together, and deducing the implied root
  element hash as a result.
  Finally we check the implied root element hash against the known
  root element hash.
- This imposes constraints on what protobufs are valid merkle proofs.
  The new routine won't accept proofs in arbitrary order, they must
  be in the order that makes validation easy. It is okay to push
  work from the server to the clients in this way.
- This also changes the merkle proof generation code to follow this
  format. It turns out that generating the proofs in this order is
  quite easy as well -- the only reason they could end up in a wierd
  order is because of the hashmap or because of sorting ambiguities,
  for example, every proof should actually contain two leaves.

Not all the tests have been fixed yet, but this draft shows the basic idea.

One thing that we could do that might make this simpler still is,
in the TxOutMembershipProof, add a separate "leaf hash" field corresponding
to the hash of the leaf at `proof.index`, and don't make a separate membership
element for this leaf. This will make some types of invalid merkle proofs
unrepresentable, and so reduce the number of error conditions of the
validation routine, essentially pushing some of the error handling
into the protobuf layer. I tried to avoid changing any proto types
in this commit though.