Remove unnecessary mutex around ByzantineLedger
The ByzantineLedger API is thread-safe, but a mutex was being used
in order to resolve a circular dependency issue.

The three major components of consensus service:
- `consensus_msgs_from_network` (background work queue taking messages)
- `broadcaster` (sends messages to our peers)
- `byzantine ledger` (processes messages and oversees SCP progress)

have needs to send messages in both directions, and so require handles
to eachother. That is, the two parts that are sending messages, sometimes
need to ask `byzantine ledger` about the state, but also, ByzantineLedger
needs to be able to push messages to the broadcaster etc.

Even though there are circular dependencies, one of them has to be
initialized last. This means that the handles passed to those that
are initialized earlier can't be `Arc<ByzantineLedger>`, there has to
be some kind of additional wrapper between `Arc` and `ByzantineLedger`
that will allow us to initialize the `ByzantineLedger` after passing
the proxy to the components that are initialized earlier.

Before this commit, that wrapper was `Mutex`. But this serializes all
interactions with the ByzantineLedger, even those that would be thread-safe.

Since the only reason mutability is needed is to permit later initialization
of the ledger, this commit changes it to be `Arc<OnceCell<ByzantineLedger>>`.
`OnceCell` is a thread-safe shared-value which can be initialized exactly once.
This avoids all the contention problems of the Mutex.

This also changes most of the handles from `Arc` to `Weak`. By making
them `Weak` we avoid circular dependencies which could prevent us
from releasing resources.