Remove unnecessary mutex around peer_keepalive
This follows up on the commit that removes the mutex around
byzantine ledger. The peer_keepalive mutex is only there to allow
stopping the service. We change this from `Arc<Mutex<PeerKeepalive>`
to `Option<Arc<PeerKeepalive>>`, and replace to call to `lock().stop()`
in the stop function with `= None` which simply drops the value.

We also make the other Arc's that are holding this, in lambda functions,
into weak references. This prevents strong reference cycles which would
prevent the drop from happening and cause a resource leak.

We also put a code comment about why not to remove the mutex around
threaded broadcaster.