Don't tear-down and copy the entire Tx cache
One of the main jobs of the TxManager is to keep track of what
transcations have been validated or partially validated by the enclave,
and keep a cache mapping their hashes to the data that untrusted
has about them.

Any part of the system that needs to know something about a particular
Tx, or add new Tx's, or purge old Tx's, or form blocks, needs to talk
to this cache object. So ideally, this cache object would be like a
thread-safe map, similar to Java's concurrent hashmap, so that we
don't bottleneck essentially all the threads in the server on this.

We haven't gotten to the point of using a concurrent hashmap for this,
in part because there are several implementations of varying quality
and choosing one is not straightforward.

We also don't have infrastructure for measuring contention of these
various locks, so we don't know that it's worth it. Right now we don't
have a good way to measure perf improvements at all,
other than just slamming the network. Usually you should find or build
such tools to measure contention at individual locks before making
changes that might or might not actually improve perfs.

Before we get to that point, there are simple things we can do to
that are obvious, strict perf improvements that don't require
measurement to validate. These changes will also help make sure that
nothing holds the lock for an unnecessarily long amount of time.

IMO before we start to do measurements of contention on individual
locks we should do a once-over making changes like this.

----

In this particular commit, we change the `remove_expired` function,
which previously was tearing down the entire `cache` and separating
it into two caches. Then all the entries in the expired section
get copied again, and the old `self.cache` variable gets overwritten
with the new value.

This is bad for a few reasons:
- It copies everything in the cache unnecessarily.
  The size of a mobilecoin transaction with 11 ring inputs is at least
  10-20k ballpark, just considering the size of merkle proofs.
  If you want to do 50 transactions per second, that's 250 transactions
  per block. If the tombstone block setting is like 10 or 20 blocks into
  the future, that's several thousand transactions.
  So copying the entire tx cache is like, ballpark, copying 100 megabytes at
  least, and ten times this if the tombstone block value is like 100 blocks
  into the future.
  And we are doing this every time we call `remove_expired`, which is
  every block.
  Arguably, a better way to do this is to use a priority queue based
  on the tombstone block, so then we don't have to scan the entire cache
  once per block. But that is a larger change that we can do later.
- Tearing down the cache means that we don't get to reuse the dynamic
  memory allocation for it, so we have to reallocate 100's of megabytes
  every time instead.
- Because the cache object is moving around randomly in memory every
  time we finish a block, the memory hierarchy cannot hope to properly
  cache it, so every thread is likely to experience cache misses
  whenever they try to talk to it. By trying to keep it in one place,
  we can avoid thrashing the caches.