Removes sgx_backtrace and sgx_debug edl interfaces in consensus_enclave edl
The sgx_backtrace thing was an attempt to port Baidu's sgx backtrace
functionality so that it would work without making system calls.
We got it working in sim mode but never actually figured out about offsets
in SGX HW mode.
Since then we became worried that it's a security hole and turned it off,
we are building in panic=abort mode for sgx, so nothing is actually using
this OCALL.

We definitly aren't going to fix this before launch, it would need major
redesign. So there's no reason for it to be in the consensus edl at all.
Probably we should go further and just delete the sgx/backtrace crate entirely.
I'm putting up this smaller commit just for discussion first.