MCC-1862 Consolidate mc-sgx-libc-types with mc-sgx-types, confirm errno values
The mc-sgx-libc-types contains a bunch of unused types that are copied
from baidu, and some duplicated with mc-sgx-types, which I think is
our preferred thing now.

The `mc-sgx-types` crate is defining a `SysError` type which we are
using, but it isn't providing the values of `SysError` which is wierd.

This fixes it by putting the errno values under `libc::SysError` in
the `mc-sgx-types` crate. Then it makes the `sgx-sync` crate, which
was getting `SysError` from here, use the errno values also from there.
Then it deletes the unused `mc-sgx-libc-types` crate and deletes
documentation about it.