Fix a minor issue in mc-util-keyfile crate (#486)
The keyfile crate writes default keyfiles and has a function
to read default keyfiles from a directory. However, due to a subtle
sorting issue, the i'th file in the returned vector is not
`account_keys_{i}.json`. This fixes that.

If this was not already causing problems in tests, then I don't think
it will cause a problem to commit this, and it will help the fog
conformance tests which are being developed.