Make a stderr configuration of the logger module, and use it in tests
Sometimes, it isn't acceptable to log on stdout, because a CLI tool
may be required to return structured output on stdout.

Typically, the stderr channel is reserved for logging in unix, and
this is what glog does by default. I'm not sure why we log to stdout.

To avoid breaking any of our partners who may be depending on the stdout
behavior, this creates a configuration for stderr, but keeps the defaults
with the current behavior.