MCC-710 Final changes regarded FullyValidatedFogPubkey and TransactionBuilder
There are a few problems with the TransactionBuilder API in regards
to fog ingest public keys.

(1) It makes the API user match public addresses with ingest public keys.
If they do this wrong, the money will not be found by the recipient.
(2) It doesn't enforce that the ingest public keys that are used are
fully validated in a type-safe way.
(3) It doesn't enforce the rule that the tombstone block of the transaction
may not exceed the pubkey_expiry of any of the fog public keys that used.

This change makes the TransactionBuilder take a set of FullyValidatedPublicKeys
object when building transactions. This object can be populated by taking
a FogPubkeyResolver and passing that, with each recipient PublicAddress,
to the object, which caches them for a short period of time.
The cache is consumed by the TransactionBuilder, reflecting the idea
that these should be fetched again with each transaction.

This caching also automatically minimizes network traffic since we
skip the second request to fog report server if two recipients have
the same fog url and report id.

This is a good idea because it turns out that in testnet block 42,
we have all zeroes for fog hints, which suggests that some client
in the past was doing fog hints wrong.

The present change was proposed months ago to try to make it harder
to use the transaction builder wrongly with respect to fog, and
hopefully reduce the number of bugs that lead to lost/missing transactions
or bad fog hints, that are hard debug due to use of encryption.