Add rpc_precondition_error, and make these use Display instead of Debug
In another PR we found use for the `FAILED_PRECONDITION` status code
of grpc.

While touching this code, we thought it would be a good idea to make
these conversion functions (which convert Error objects to appropriate
RpcStatus objects from grpcio crate), report the Error objects
using `Display` instead of `Debug`, because all of our Error objects
implement `Display` and it's usually more readable to a human than
the `Debug` formatting. This also makes the "context" thing use `Display`
as well instead of `ToString`, since in reality almost all `ToString`
implementations come from `Display` because of the blanket implementation.