Replace Aes128 with Aes256 everywhere that was using it (#725)
* Replace Aes128 with Aes256 everywhere that was using it

During review, we noticed that sometimes we select Aes128 and
sometimes Aes256. The mc-noise crate is using Aes256 to create
attested connections to enclaves. The mc-crypto-box is using Aes128,
and the message-cipher was also using Aes128.

Using two kinds of AES, especially in attested connections and
in cryptobox, means that an embedded device that wants to construct
mobilecoin transactions needs to support both forms of AES. Usually
it's better to just use one primitive everywhere to reduce code
foot print. AES256 is not appreciably slower than AES128, so it is
fine to replace Aes128 with Aes256 everywhere.

The change is likely going to change the MRENCLAVE value of consensus,
if only because of the message cipher change.
We also noticed that there is an unused file in crypto/box/src.
The real implementation is `hkdf_box.rs` and this is generic over
the different primitives. The old file may have gotten committed
again by accident somehow, I'm not sure.

Note: This is a breaking change to crypto-box, but since Fog has
not been deployed in production anywhere, this is fine.

* update some test vectors

the test vectors change because the data values that are being
generated are changing. in the future it would be better to make
the input to the test deterministic as well and not dependent on
transaction crate stuff.

* really fix test vectors