Add slam tool to consensus server repo
The slam tool is a load testing tool for the consensus servers.
Generally, test code should be versioned with the code under test.

This moves the slam tool to the same repo as the consensus servers,
and will make it easier to run this test in CD and more easily use
it to validate changes to the consensus servers.