Expose the `get_value_mask` function of `transaction_core::amount`
This slightly refactors the `get_value_mask` function to return the u64
mask instead of a Ristretto scalar. This is because the u64 is what
all of the callers need and use, so there is less code duplication
if the function actually returns that.

This also makes the `get_value_mask` function exposed as part of the
public transaction-core API. This is because it is needed in fog,
to support a version of TxOut's where we don't store the compressed
commitment in Fog, and instead the user reconstructs that from the
other data in the TxOut.