WIP recoverable transaction history, encrypted memos on TxOut's
This adds an e_memo field to TxOut, which is an opaque memo field
encrypted in a standard way and interprettable in a standard way
by all clients. The format is designed to be extensible to meet
future needs, while maintaining output uniformity for privacy.

The functionality which allows creating "MemoPayload" data and
encrypting it for the TxOut is added to the transaction-core crate.

The functionality which allows creating high-level memo types,
and interpretting memo payloads as a known high-level type, is added
to the transaction-std crate. This is also integrated with the
transaction-builder crate so that it mostly does the right thing,
but the defaults are still for no information disclosure via memos.

Due to properties of mc-crypto-digestible, adding new optional fields
like this does not change the hash of old transactions.

This also adds some functionality to `mc-account-keys`:
- Change subaddress is standardized, and method is added to compute it
- A standard `AddressHash` of a public address is created using the
  `mc-crypto-digestible` crate. This hash is used in memos.

Still to do:
- [ ] test coverage on memo creation and validation apis
- [ ] test coverage on memo type round trips
- [ ] test coverage on interpretting memos from transaction builder
- [ ] further discussion around transaction builder API changes
- [ ] increment block version number
- [ ] add option to mobilecoind for recoverable transaction history
      feature, and test this functionality (is this desired?)
- [ ] make mobilecoind use change subaddress for change, setting
      aside issues of memos (is this desired?)
- [ ] update fog to carry the e_memo through to fog-view, and tests
- [ ] update fog sample paykit to support creating and interpretting
      recoverable transaction history, and tests of this
- [ ] update fog test client to exercise this functionality, make
      sure this is tested as part of CD