fix documentation for onetime key vs target key
It is very confusing to have two names for the same value used
in source code. In TxOut's we have a value called "target_key",
but in some functions in `onetime_keys.rs` this is called the
`onetime_public_key`.

"One-time public key" is not a good technical term for a key,
because many keys in cryptography are one-time use, and this name
does not disambiguate the key from any other one-time use key.

This commit unifies the source code and code comments around the
term "target_key" which is what we are already using in the TxOut
structure.