`git mv fog mc-fog`, and add readme
This is needed because some crates like `fog/api` have paths which
will be the same as a crate from `fog` repo.

We can work on reconciling crates like that after we have merged the
repos and tests are passing, but we don't want to do any code changes
in this commit.