Make Encrypted memo changes releasable
In testing, we discovered that old clients are unable to submit
transactions to new servers, even if the MRENCLAVE value is up to date.
This is because the entire TxPrefix is under the ring signature in the
transaction math, and the TxPrefix is different whether or not a memo
is there. An old server has protobufs that don't include the memo field,
so they compute the hash of the TxPrefix differently from clients that
included an empty memo.

To make this go, we decided to release this change in two steps.
First, we make versions of the servers that tolerate the presence of
the memo and compute the TxPrefix hash in the same way as the clients,
but also accept the state of affairs that there are no memos.

Existing clients can thus continue to work with this server simply by
updating the MRENCLAVE value that they have hard coded.

In a subsequent server release, the memos will become mandatory.
In the transition period before that release, Clients will all start
using the memos, and reading from them, expecting to find data there
(but tolerating the absence of that data, so that these clients work
in the period before this secondary server release, during which
memos are optional.)

The changes are:
* Don't require every TxOut to have a memo during validation
  (but keep that functionality under test, and refer to a ticket to
  bring it back.)
* MemoFn in the TxOut::new, MemoBuilder, and TransactionBuilder pathways
  returns `Result<Option<MemoPayload>,...>` rather than `Result<MemoPayload, ...>`
  to allow that memo builders may omit the memo.
  (This modificaiton will be reverted when memos are made mandatory.)
* A `NoMemoBuilder` is added in the `transaction-std` crate, which always
  produces `None`. This is intended to be used during the transationary
  period when the servers have not yet started accepting memos.
* Make android-bindings, libmobilecoin, mobilecoind, use the `NoMemoBuilder`.
  They can be switched to `EmptyMemoBuilder` or `RTHMemoBuilder` as necessary
  during the next release.